requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic[email]==2.5.0

# Environment and configuration
python-dotenv==1.0.0

# Database
pymongo==4.6.0

# Vector embeddings and AI
cohere==4.37
groq==0.4.1
numpy==1.26.2

# PDF processing
PyPDF2==3.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Optional development dependencies
pytest==7.4.3
httpx==0.25.2